move classes to a new package
