+++ app/src/main/java/com/android/dev/engineer/kotlin/compose/data/use_case/initial_route/GetInitialRouteUseCaseImpl.kt
@@ -2,7 +2,7 @@ package com.android.dev.engineer.kotlin.compose.data.use_case.initial_route
 
 import com.android.dev.engineer.kotlin.compose.data.di.IoDispatcher
 import com.android.dev.engineer.kotlin.compose.data.domain.MainNavGraph
-import com.android.dev.engineer.kotlin.compose.data.local.AppDataStore
+import com.android.dev.engineer.kotlin.compose.data.datastore.app.AppDataStore
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
