+++ app/src/main/java/com/android/dev/engineer/kotlin/compose/data/use_case/mark_intro/MarkIntroCompleteUseCaseImpl.kt
@@ -1,7 +1,7 @@
 package com.android.dev.engineer.kotlin.compose.data.use_case.mark_intro
 
 import com.android.dev.engineer.kotlin.compose.data.di.IoDispatcher
-import com.android.dev.engineer.kotlin.compose.data.local.AppDataStore
+import com.android.dev.engineer.kotlin.compose.data.datastore.app.AppDataStore
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
